pyproject.toml
[project]
name = "utilities-payment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
VIRTUAL_ENV= ".venv"
dependencies = [
    "celery>=5.6.3",
    "django>=5.2.12",
    "django-celery-beat>=2.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "flower>=2.0.1",
]
